Заняття 43
Додано модуль пагінації

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -152,6 +153,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updated_price;
             product.description = updated_description;
             if (updated_imageUrl) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = updated_imageUrl.path;
             }
             return product.save().then(result => {
@@ -168,7 +170,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(result => {
             console.log('Product deleted!');
             res.redirect('/admin/products');
